added Rag History tab on Assessment tab using assessment history converted as a component
