+++ src/Server.UI/Pages/Assessment/AssessmentComponents/AssessmentHistory.razor
@@ -1,0 +1,33 @@
+@using Cfo.Cats.Application.Features.Assessments.DTOs
+@inherits CatsComponentBase
+
+@if (ParticipantAssessments.Any() == true)
+{
+<MudDataGrid Items="@ParticipantAssessments" Striped="true">
+    <Columns>
+        <TemplateColumn Title="Completed" Sortable="true" SortBy="x => x.Completed">
+            <CellTemplate>
+                <MudStack>
+                    <MudText Typo="Typo.body2">@context.Item.Completed!.Value.ToShortDateString()</MudText>
+                    <MudText Typo="Typo.body2">@($"{GetSinceConsentDate(context.Item.Completed)} (since consent)")</MudText>
+                </MudStack>
+            </CellTemplate>
+        </TemplateColumn>
+        <PropertyColumn Property="x => x.LocationName" Title="Location"></PropertyColumn>
+        <TemplateColumn Title="RAG Bar">
+            <CellTemplate>
+                <Cfo.Cats.Server.UI.Pages.Participants.Components.RagBar Model="@context.Item" />
+            </CellTemplate>
+        </TemplateColumn>
+    </Columns>
+    <PagerContent>
+        <MudDataGridPager T="ParticipantAssessmentDto" />
+    </PagerContent>
+</MudDataGrid>
+}
+else
+{
+    <MudAlert>
+        <MudAlert Severity="Severity.Info" Variant="Variant.Outlined" Square="true" Class="my-2">No history found.</MudAlert>
+    </MudAlert>
+}
