--- src/Server.UI/Pages/Assessment/AssessmentHistory.razor
@@ -1,60 +1,0 @@
-@using Cfo.Cats.Application.Features.Participants.Commands
-@using Cfo.Cats.Application.Features.Participants.Queries
-
-@using Cfo.Cats.Application.Features.Assessments.DTOs
-@using Cfo.Cats.Application.Features.Assessments.Queries
-@using Cfo.Cats.Server.UI.Pages.Participants.Components
-
-@inherits CatsComponentBase
-
-@page "/pages/participants/{Upci}/assessmenthistory"
-
-
-@if (_participantAssessments.Any() == true && _participant is not null)
-{
-    <MudPaper Class="pa-2 ma-2" Elevation="3">
-
-        <MudGrid Class="pa-1" Justify="Justify.FlexStart">
-            <MudItem xs="12">
-
-                <MudStack Row="true" Spacing="2" Reverse="false" Breakpoint="Breakpoint.SmAndDown" AlignItems="AlignItems.Start" Justify="Justify.FlexStart" Wrap="Wrap.Wrap"  >
-                    <MudText Typo="Typo.subtitle1" Color="Color.Info"><strong>Participant</strong></MudText>
-                    <MudText Typo="Typo.subtitle1">@_participant!.FullName</MudText>
-                    <MudDivider Vertical="true" FlexItem="true"/>
-                    <MudText Typo="Typo.subtitle1" Color="Color.Info"><strong>Ref</strong></MudText>
-                    <MudText Typo="Typo.subtitle1">@_participant!.Id</MudText>
-                    <MudDivider Vertical="true" FlexItem="true"/>
-                    <MudText Typo="Typo.subtitle1" Color="Color.Info"><strong>Status</strong></MudText>
-                    <MudText Typo="Typo.subtitle1">@_participant!.EnrolmentStatus!.Name</MudText>
-                    <MudDivider Vertical="true" FlexItem="true"/>
-                </MudStack>
-
-            </MudItem>
-        </MudGrid>
-
-        <MudGrid Class="pa-1" Justify="Justify.FlexStart">
-        @foreach (var participantAssessmentDto in _participantAssessments)
-        {
-            <MudItem md="12">
-                <MudStack Row="true" Spacing="2" Reverse="false" Breakpoint="Breakpoint.SmAndDown" AlignItems="AlignItems.Start" Justify="Justify.FlexStart" Wrap="Wrap.Wrap"  >
-                    <MudText Typo="Typo.subtitle1" Color="Color.Info"><strong>Completed</strong></MudText>
-                        <MudText Typo="Typo.subtitle1">@participantAssessmentDto.Completed!.Value.ToString("dd/MM/yyyy")</MudText>
-                    <MudDivider Vertical="true" FlexItem="true"/>
-                    <MudText Typo="Typo.subtitle1">
-                        @GetSinceConsentDate(participantAssessmentDto.Completed)
-                    </MudText >
-                    <MudText Typo="Typo.subtitle1" Color="Color.Info">since participant consented</MudText>
-
-                </MudStack>
-                <RagBar Model="@participantAssessmentDto" />
-                <MudStack Row="true" Spacing="2" Reverse="false" Breakpoint="Breakpoint.SmAndDown" AlignItems="AlignItems.Start" Justify="Justify.FlexStart" Wrap="Wrap.Wrap">
-                    <MudText Typo="Typo.subtitle1" Color="Color.Info"><strong>Assessment Location</strong></MudText>
-                    <MudText Typo="Typo.subtitle1">@participantAssessmentDto.LocationName</MudText>
-                </MudStack>
-                    <MudDivider Class="my-3" />
-            </MudItem>
-            }
-        </MudGrid>
-    </MudPaper>
-}
-
