+++ src/Server.UI/Pages/Participants/Components/CaseAssessment.razor
@@ -1,5 +1,4 @@
 @using Cfo.Cats.Application.Features.Assessments.DTOs
-@using Cfo.Cats.Application.Features.Assessments.Queries
 @using Cfo.Cats.Domain.Common.Enums
 
 @inherits CatsComponentBase
@@ -24,110 +23,73 @@
     
 </style>
 
-@if (_notFound)
-{
-    <MudAlert>
-        <MudAlert Severity="Severity.Info" Variant="Variant.Outlined" Square="true" Class="my-2">No assessment found.</MudAlert>
-    </MudAlert>
-}
 
-
-@if (_model is not null)
-{
-    <MudExpansionPanels Class="pa-3 m-3" MultiExpansion="true">
-        
-        @foreach (var pathway in _model.Pathways)
-        {
-            <MudExpansionPanel>
-                <TitleContent>
-                    <div class="d-flex">
-                        <MudIcon Icon="@pathway.Icon" Color="Color.Primary" class="mr-3"/>
-                        <MudText>@pathway.Title</MudText>
-                    </div>
-                </TitleContent>
-                <ChildContent>
-                    <MudGrid Class="mb-4 px-10">
-
-                        @foreach (var question in pathway.Questions())
+    <MudTabs Outlined="true" Border="true" Class="mt-2" >
+        <MudTabPanel Text="Latest">
+            <MudExpansionPanels Class="pa-3 m-3" MultiExpansion="true">
+                @if (_notFoundAssessment)
+                {
+                    <MudAlert>
+                        <MudAlert Severity="Severity.Info" Variant="Variant.Outlined" Square="true" Class="my-2">No assessment found.</MudAlert>
+                    </MudAlert>
+                }
+                @if (_model is not null)
+                    {
+                        @foreach (var pathway in _model.Pathways)
                         {
+                            <MudExpansionPanel>
+                                <TitleContent>
+                                    <div class="d-flex">
+                                        <MudIcon Icon="@pathway.Icon" Color="Color.Primary" class="mr-3" />
+                                        <MudText>@pathway.Title</MudText>
+                                    </div>
+                                </TitleContent>
+                                <ChildContent>
+                                    <MudGrid Class="mb-4 px-10">
 
-                        <MudItem xs="4">
-                            <Cfo.Cats.Server.UI.Pages.Assessment.AssessmentComponents.AssessmentQuestion Question="@question.Question" HelperText="@question.OtherInformation" />
-                        </MudItem>
+                                        @foreach (var question in pathway.Questions())
+                                        {
 
-                        <MudItem xs="8">
-                            @if (question is SingleChoiceQuestion atq)
-                            {
-                            <MudToggleGroup T="string" Class="readonly-toggle" SelectionMode="SelectionMode.SingleSelection" @bind-Value="@atq.Answer" CheckMark FixedContent Color="Color.Primary">
-                                @foreach (var item in atq.Options)
-                                {
-                                    <MudToggleItem Value="@item" Text="@item" />
-                                }
-                            </MudToggleGroup>
+                                            <MudItem xs="4">
+                                                <Cfo.Cats.Server.UI.Pages.Assessment.AssessmentComponents.AssessmentQuestion Question="@question.Question" HelperText="@question.OtherInformation" />
+                                            </MudItem>
 
-                            }
+                                            <MudItem xs="8">
+                                                @if (question is SingleChoiceQuestion atq)
+                                                {
+                                                    <MudToggleGroup T="string" Class="readonly-toggle" SelectionMode="SelectionMode.SingleSelection" @bind-Value="@atq.Answer" CheckMark FixedContent Color="Color.Primary">
+                                                        @foreach (var item in atq.Options)
+                                                        {
+                                                            <MudToggleItem Value="@item" Text="@item" />
+                                                        }
+                                                    </MudToggleGroup>
 
-                            @if (question is MultipleChoiceQuestion amcq)
-                            {
-                                <MudToggleGroup T="string" Class="two-columns readonly-toggle" Vertical SelectionMode="SelectionMode.MultiSelection" @bind-Values="@amcq.Answers" CheckMark Color="Color.Primary" Outlined="false" Delimiters="false">
-                                    @foreach (var item in amcq.Options)
-                                    {
-                                        <MudToggleItem Class="toggle-item" Value="@item" Text="@item" UnselectedIcon="@Icons.Material.Filled.CheckBoxOutlineBlank" SelectedIcon="@Icons.Material.Filled.CheckBox" />
-                                    }
-                                </MudToggleGroup>
-                            }
-                        </MudItem>
-                        <MudDivider />
+                                                }
+
+                                                @if (question is MultipleChoiceQuestion amcq)
+                                                {
+                                                    <MudToggleGroup T="string" Class="two-columns readonly-toggle" Vertical SelectionMode="SelectionMode.MultiSelection" @bind-Values="@amcq.Answers" CheckMark Color="Color.Primary" Outlined="false" Delimiters="false">
+                                                        @foreach (var item in amcq.Options)
+                                                        {
+                                                            <MudToggleItem Class="toggle-item" Value="@item" Text="@item" UnselectedIcon="@Icons.Material.Filled.CheckBoxOutlineBlank" SelectedIcon="@Icons.Material.Filled.CheckBox" />
+                                                        }
+                                                    </MudToggleGroup>
+                                                }
+                                            </MudItem>
+                                            <MudDivider />
+                                        }
+                                    </MudGrid>
+                                </ChildContent>
+                            </MudExpansionPanel>
                         }
-                    </MudGrid>
-                </ChildContent>
-            </MudExpansionPanel>
-        }
+                    }
+            </MudExpansionPanels>
+        </MudTabPanel>
+        <MudTabPanel Text="RAG History">
+        <Cfo.Cats.Server.UI.Pages.Assessment.AssessmentComponents.AssessmentHistory ParticipantAssessments="_participantAssessments" ConsentDate="_consentDate" />
+        </MudTabPanel>
+    </MudTabs>
 
-    </MudExpansionPanels>
-}
 
 
 
-@code {
-
-    private Assessment? _model;
-    private bool _notFound = false;
-
-    [Parameter]
-    [EditorRequired]
-    public string ParticipantId { get; set; } = default!;
-
-    protected override async Task OnInitializedAsync()
-    {
-        if(_model is not null)
-        {
-            return;
-        }
-
-        try
-        {
-            var result = await GetNewMediator().Send(new GetAssessment.Query()
-            {
-                ParticipantId = ParticipantId
-            });
-
-            if(result.Succeeded)
-            {
-                _model = result.Data;
-            }
-
-        }
-        finally
-        {
-            _notFound = _model is null;
-        }
-
-    }
-
-    protected override async Task OnAfterRenderAsync(bool firstRender)
-    {
-        await JS.InvokeVoidAsync("removeInlineStyle", ".two-columns");
-    }
-
-}
