+++ src/Server.UI/Pages/Participants/Components/CaseSummary.razor.cs
@@ -128,11 +128,6 @@ public partial class CaseSummary
         Navigation.NavigateTo($"/pages/participants/{ParticipantSummaryDto.Id}/assessment/{_latestAssessment!.AssessmentId}");
     }
 
-    public void ViewAssessmentHistory()
-    {
-        Navigation.NavigateTo($"/pages/participants/{ParticipantSummaryDto.Id}/assessmenthistory");
-    }
-
     /// <summary>
     /// If true, indicates we are creating our first ever assessment. 
     /// </summary>
@@ -164,8 +159,6 @@ public partial class CaseSummary
         } &&
                ParticipantSummaryDto.EnrolmentStatus.SupportsReassessment();
     }
-
-    private bool IsAssessmentHistoryAvailable() => ParticipantSummaryDto.Assessments.Any(a => a.Completed.HasValue);
 
     private bool HasRiskInformation() => ParticipantSummaryDto.LatestRisk is not null;
     private bool CanAddRiskInformation() => HasRiskInformation() is false;
